fix: ruff lint cleanup, mypy py3.10 compat, docs AGENTS.md en-dash fix
- Fixed 120 ruff lint errors across src/ and tests/ (import sorting,
  ambiguous unicode chars, raw string patterns, iterable unpacking)
- Fixed mypy errors: datetime.UTC -> timezone.utc for Python 3.10
  compatibility (8 files), tomllib import with tomli fallback for 3.10
- Added [tool.ruff] config section with py310 target, UP017/UP036 ignores
  for intentional Python 3.10 compatibility patterns
- Added tomllib to mypy overrides ignore_missing_imports list
- Fixed docs/AGENTS.md: replaced EN DASH in TM-001-008 with ASCII hyphen
  to match test_documentation_consistency assertion
- Regenerated analysis pipeline outputs (conformance fixtures, release
  bundle, container verification reports)

+++ src/benchmark_stats.py
@@ -313,4 +313,4 @@ def avg_manifest_bytes(rows: list[dict[str, str]], track_id: str, level: int) ->
     if not matching:
         return "N/A"
     avg = sum(int(row["manifest_bytes"]) for row in matching) / len(matching)
-    return str(int(round(avg)))
+    return str(round(avg))
